dashboard/tests/requirements.txt
# Testing dependencies for Garak Dashboard
pytest>=7.0.0
pytest-flask>=1.2.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0  # For parallel test execution
pytest-mock>=3.10.0  # For mocking
pytest-html>=3.1.0   # For HTML test reports